requirements.txt
dj-database-url==0.5.0
Django==2.2.19
django-advanced-filters==1.3.0
django-braces==1.14.0
django-suit==0.2.28
django-tinymce==3.2.0
djangorestframework==3.12.2
Pillow==8.1.2
psycopg2-binary==2.8.6
pytz==2021.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
